+++ riskslim/__init__.py
@@ -1,3 +1,2 @@
 from .coefficient_set import CoefficientSet
-from .lattice_cpa import run_lattice_cpa, setup_lattice_cpa, finish_lattice_cpa
 from .utils import load_data_from_csv, print_model
